parseInt -> parseFloat

+++ server/cache/rating.js
@@ -47,7 +47,7 @@ async function insertRating(userID, productID, rate) {
         if (await isRatingYet(userID, productID)) throw new Error('User rated this product')
         await pool.query(queryString, [userID, productID, rate])
         const [rows] = await pool.query(queryString1, [productID])
-        product.updateRating(parseInt(rows[0].rate))
+        product.updateRating(parseFloat(rows[0].rate))
         product.updateRatingCount()
     } catch (error) {
         throw new Error(error.message)
@@ -67,7 +67,7 @@ async function updateRating(userID, productID, rate) {
         if (!await isRatingYet(userID, productID)) throw new Error('User didn\'t rated this product')
         await pool.query(queryString, [rate, userID, productID])
         const [rows] = await pool.query(queryString1, [productID])
-        product.updateRating(parseInt(rows[0].rate))
+        product.updateRating(parseFloat(rows[0].rate))
     } catch (error) {
         throw new Error(error.message)
     }
